test: create tests for LINQ methods
